contracts removed MediaType and FileTyped & , gloabl metadata accepts MediaType
