+++ src/camouchat_core/contracts/__init__.py
@@ -8,7 +8,7 @@ from .chat import ChatProtocol
 from .chat_processor import ChatProcessorProtocol
 from .interaction_controller import InteractionControllerProtocol
 from .login import LoginProtocol
-from .media_controller import FileTyped, MediaControllerProtocol, MediaType
+from .media_controller import MediaControllerProtocol
 from .message import MessageProtocol
 from .message_processor import MessageProcessorProtocol
 from .storage import StorageProtocol
@@ -24,6 +24,4 @@ __all__ = [
     "UiConfigProtocol",
     "StorageProtocol",
     "LoginProtocol",
-    "MediaType",
-    "FileTyped",
 ]
