+++ src/camouchat_core/global_metadata/media_type.py
@@ -1,0 +1,11 @@
+from enum import StrEnum
+
+
+class MediaType(StrEnum):
+    """Supported media types for upload across all platforms."""
+
+    TEXT = "text"
+    IMAGE = "image"
+    VIDEO = "video"
+    AUDIO = "audio"
+    DOCUMENT = "document"
