+++ src/camouchat_core/global_metadata/msg_type.py
@@ -1,5 +1,6 @@
 from enum import StrEnum
 
+
 class MessageType(StrEnum):
     TEXT = "text"
     IMAGE = "image"
